refactor(tsm1): Allow race-free access to cache
This commit adds a new API to `Cache` to fix data races
with the existing `TagKeys` and `TagValues` APIs.

The `ApplyEntryFn` calls the provided callback function,
which passes a cache `entry`. Accessing the `values` field
is a data race, as it may be mutated by other goroutines.

`entry` provides new APIs, `Len`, `Contains` and `MaxTime`
that are safe.

Additionally, `Cache` now provides `ApplyEntryFnForKey`
to call `f` for the specified key. During the callback,
a read lock is held for the `entry`, safely allowing access
to `values`